fix some comments; move vit flags to sep file
